requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pdfplumber==0.10.0
pytesseract==0.3.10
pillow==10.1.0
beautifulsoup4==4.12.2
requests==2.31.0
duckduckgo-search==3.9.6
selenium==4.15.2
webdriver-manager==4.0.1
python-docx==1.1.0
openpyxl==3.1.2
ollama==0.1.8
aiohttp==3.9.1
websockets==12.0
jinja2==3.1.2
aiofiles==23.2.1
pandas==2.1.3
numpy==1.26.2
pdf2image==1.16.3
python-dateutil==2.8.2
